Move vector.h to shared location to resolve Linux compilation issues
- Created src/game/shared/hl/vector.h for cross-platform access
- Updated include paths in AngelScript files to use shared location
- Updated include paths in client and server files
- Removed old vector.h from server/hl directory

This resolves Linux compilation issues by providing a consistent
shared vector.h location that all components can reference without
complex relative paths.
