+++ src/game/shared/ms/angelscript/ASCoreTypes.cpp
@@ -14,7 +14,7 @@
 // Include vector.h directly to avoid mathlib.h macro conflicts
 // We need the vec_t typedef but not the DotProduct macro
 typedef float vec_t;
-#include "../../../../server/hl/vector.h"
+#include "hl/vector.h"
 
 // Include MSLogger for consistent logging
 #include "mslogger.h"
