+++ src/game/shared/ms/angelscript/ASEngineEventManager.cpp
@@ -12,7 +12,7 @@
 // Include vector.h directly to avoid mathlib.h macro conflicts
 // We need the vec_t typedef but not the DotProduct macro
 typedef float vec_t;
-#include "../../../../server/hl/vector.h"
+#include "hl/vector.h"
 
 // Static instance
 ASEngineEventManager* ASEngineEventManager::s_pInstance = nullptr;
